robot-controller/requirements.txt
Flask
requests
numpy
Pillow